+++ frontend/src/components/layouts/Layout/index.tsx
@@ -2,7 +2,7 @@ import React, { ReactNode, useMemo } from 'react';
 import { useRouter } from 'next/router';
 import { signOut, useSession } from 'next-auth/react';
 
-import LoadingPage from '@/components/loadings/LoadingPage';
+import LoadingPage from '@/components/Primitives/Loading/Page';
 import { Sidebar } from '@/components/Sidebar';
 import { REFRESH_TOKEN_ERROR } from '@/utils/constants';
 import { BOARDS_ROUTE, DASHBOARD_ROUTE, TEAMS_ROUTE, USERS_ROUTE } from '@/utils/routes';
