+++ frontend/src/pages/boards/[boardId].tsx
@@ -16,7 +16,7 @@ import AlertMergeIntoMain from '@/components/Board/SplitBoard/AlertMergeIntoMain
 import BoardHeader from '@/components/Board/SplitBoard/Header';
 import Timer from '@/components/Board/Timer';
 import Icon from '@/components/icons/Icon';
-import LoadingPage from '@/components/loadings/LoadingPage';
+import LoadingPage from '@/components/Primitives/Loading/Page';
 import AlertBox from '@/components/Primitives/AlertBox';
 import Button from '@/components/Primitives/Button';
 import Flex from '@/components/Primitives/Flex';
