+++ frontend/src/pages/boards/index.tsx
@@ -5,7 +5,7 @@ import MyBoards from '@/components/Boards/MyBoards';
 import QueryError from '@/components/Errors/QueryError';
 import requireAuthentication from '@/components/HOC/requireAuthentication';
 import Layout from '@/components/layouts/Layout';
-import LoadingPage from '@/components/loadings/LoadingPage';
+import LoadingPage from '@/components/Primitives/Loading/Page';
 import Flex from '@/components/Primitives/Flex';
 import useTeam from '@/hooks/useTeam';
 import { teamsListState } from '@/store/team/atom/team.atom';
